# openapi.yaml
openapi: 3.0.0
info:
  title: 'Medication REST Api'
  version: 1.0.0
paths:
...
  /api/loans:
    get:
      tags:
        - Loans
      summary: 'Get list of loans'
      description: 'Returns list of all loans'
      operationId: getLoansList
      parameters:
        -
          name: per_page
          in: query
          description: 'Number of items per page'
          required: false
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Loan' } }
                type: object
  '/api/clients/{clientId}/loans':
    get:
      tags:
        - Loans
      summary: "Get client's loans"
      description: 'Returns list of loans for a specific client'
      operationId: getClientLoans
      parameters:
        -
          name: clientId
          in: path
          description: 'Client id'
          required: true
          schema:
            type: integer
            format: int64
        -
          name: per_page
          in: query
          description: 'Number of items per page'
          required: false
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Loan' } }
                type: object
    post:
      tags:
        - Loans
      summary: 'Apply for a loan'
      description: 'Process a loan application for a client'
      operationId: applyForLoan
      parameters:
        -
          name: clientId
          in: path
          description: 'Client id'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Loan application data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanApplicationRequest'
      responses:
        '201':
          description: 'Loan approved'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Loan' }
                  message: { type: string, example: 'Loan application approved' }
                type: object
        '422':
          description: 'Loan application rejected'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Loan application rejected' }
                  reasons: { type: array, items: { type: string }, example: ['Credit score must be greater than 500', 'Income must be at least $1000'] }
                type: object
  '/api/clients/{clientId}/eligibility':
    get:
      tags:
        - Loans
      summary: 'Check loan eligibility'
      description: 'Check if a client is eligible for a loan without creating one'
      operationId: checkLoanEligibility
      parameters:
        -
          name: clientId
          in: path
          description: 'Client id'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Eligibility check result'
          content:
            application/json:
              schema:
                properties:
                  eligible: { type: boolean, example: true }
                  messages: { type: array, items: { type: string }, example: [] }
                type: object
        '404':
          description: 'Client not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Client not found' }
                type: object
  '/api/loans/{id}':
    get:
      tags:
        - Loans
      summary: 'Get loan information'
      description: 'Returns loan data'
      operationId: getLoanById
      parameters:
        -
          name: id
          in: path
          description: 'Loan id'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Loan' }
                type: object
        '404':
          description: 'Loan not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Loan not found' }
                type: object
components:
  schemas:
    LoanApplicationRequest:
      title: 'Loan Application Request'
      description: 'Request for applying for a loan'
      required:
        - name
        - amount
        - rate
      properties:
        name:
          description: 'Loan name/type'
          type: string
          example: 'Personal Loan'
        amount:
          description: 'Loan amount in USD'
          type: number
          format: float
          example: 1000
        rate:
          description: 'Interest rate percentage'
          type: number
          format: float
          example: 10
        start_date:
          description: 'Loan start date (optional, defaults to current date)'
          type: string
          format: date
          example: '2024-01-01'
        end_date:
          description: 'Loan end date (optional, defaults to one year after start date)'
          type: string
          format: date
          example: '2024-12-31'
      type: object
...
    Loan:
      title: Loan
      description: 'Loan model'
      properties:
        id:
          description: 'Loan ID'
          type: integer
          format: int64
          example: 1
        client_id:
          description: 'Client ID'
          type: integer
          format: int64
          example: 1
        name:
          description: 'Loan name/type'
          type: string
          example: 'Personal Loan'
        amount:
          description: 'Loan amount in USD'
          type: number
          format: float
          example: 1000
        rate:
          description: 'Interest rate percentage'
          type: number
          format: float
          example: 10
        start_date:
          description: 'Loan start date'
          type: string
          format: date
          example: '2024-01-01'
        end_date:
          description: 'Loan end date'
          type: string
          format: date
          example: '2024-12-31'
        status:
          description: 'Loan status'
          type: string
          enum:
            - pending
            - approved
            - rejected
          example: approved
        created_at:
          description: 'Creation timestamp'
          type: string
          format: date-time
        updated_at:
          description: 'Last update timestamp'
          type: string
          format: date-time
      type: object
...